app/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Script==2.0.6
graypy==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
# protobuf==3.12.1
uWSGI==2.0.17.1
six==1.15.0
Werkzeug==1.0.1
